[MOB-32] Moving some files to other folder
