+++ src/infrastructure/dtos/index.ts
@@ -1,6 +1,5 @@
 export * from './event.dto';
 export * from './event-list.dto';
-export * from './account-login.dto';
-export * from './user-profile-dto';
+export * from './user-profile.dto';
 export * from './account-logged.dto';
-export * from './pair-of-tokens';
+export * from './pair-of-tokens.dto';
