+++ src/infrastructure/http/dtos/index.ts
@@ -1,1 +1,2 @@
 export * from './create-event.dto';
+export * from './account-login.dto';
